+++ packages/datadog_flutter_plugin/android/src/test/kotlin/com/datadoghq/flutter/DatadogSdkPluginTest.kt
@@ -152,7 +152,9 @@ class DatadogSdkPluginTest {
             trackingConsent = TrackingConsent.GRANTED,
             rumConfiguration = DatadogFlutterConfiguration.RumConfiguration(
                 applicationId = applicationId,
-                sampleRate = 82.3f
+                sampleRate = 82.3f,
+                detectLongTasks = true,
+                longTaskThreshold = 0.1f
             )
         )